Move WithTypeIds into a module
